+++ Assets/Scripts/MyInterfaces/IDamageable.cs
@@ -2,11 +2,11 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    public float MaxHealth { get; set; }    
-    public float CurrentHealth { get; set; }
+    public int MaxHealth { get; set; }    
+    public int CurrentHealth { get; set; }
     public float DamageModifier { get; set; }
 
     public void RestoreFullHealth(); 
-    public void ApplyDamage(float amount);
+    public void ApplyDamage(int amount);
     public void Die();
 }
