+++ Assets/Scripts/MyInterfaces/IHitable.cs
@@ -2,12 +2,12 @@ using UnityEngine;
 
 public struct HitInfo
 {
-    public float Damage;
+    public int Damage;
     public HitSfxType HitSfx;
     public Vector3 AttackerPosition;
     public float KnockbackForce;
 
-    public HitInfo(float damage, HitSfxType hitSfx, Vector3 attackerPosition, float knockbackForce)
+    public HitInfo(int damage, HitSfxType hitSfx, Vector3 attackerPosition, float knockbackForce)
     {
         Damage = damage;
         HitSfx = hitSfx;
