+++ Assets/Scripts/MyInterfaces/IProjectile.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public interface IProjectile
 {
     public string TargetTag { get; set; }
-    public float Damage {  get; set; }
+    public int Damage {  get; set; }
     public float Speed { get; set; }
     public float LifeTime {  get; set; }
